+++ NEWS.md
@@ -7,3 +7,5 @@
 * Add `--no-rc` to `create_env_*()`, `packages_search_*()`, and `install_packages()` (v.0.0.3.9020).
 
 * Add `--clean-env` to `run_*()` functions v(0.0.3.9021).
+
+* Add `--no-rc` to `list_envs()` and `list_packages()` (v.0.0.3.9022).